requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
asyncpg==0.27.0
attrs==23.1.0
charset-normalizer==3.1.0
discord-ui==5.1.6
discord.py==2.3.1
frozenlist==1.3.3
greenlet==2.0.2
idna==3.4
multidict==6.0.4
python-dotenv==1.0.0
SQLAlchemy==2.0.17
typing_extensions==4.6.3
yarl==1.9.2
